moving create post query into own files
