+++ client/src/api/index.ts
@@ -4,4 +4,5 @@ export * from './queries/thread';
 
 // Hook exports
 export * from './hooks/usePoll';
-export * from './hooks/useThreads'; 
+export * from './hooks/useThreads';
+export * from './hooks/useCreatePost';
